Compiler.Analysis.DependencyExtraction

Description

This module contains functions to extract the names of (type) constructors and variables (including functions) used by a Haskell type, expression or declaration.

These functions are used by Compiler.Analysis.DependencyGraph to construct the dependency graph. They are also used during the conversion of function declarations to extract the type variables used in the type signature (becuase in Coq we need to declare type variables explicitly).

Types

typeDependencies :: TypeDependencies a => a -> [QName]

Extracts the names of all type variables and type constructors used in the given type expression.

This also includes the names of predefied constructors.

typeVars :: TypeDependencies a => a -> [QName]

Extracts the names of all type variables used in the given type expression.

typeVarSet :: TypeDependencies a => a -> Set QName

Like typeVars but returns a set of variable names.

typeCons :: TypeDependencies a => a -> [QName]

Extracts the names of all type constructors used in the given type expression.

This also includes the names of predefined constructors (e.g. the list and pair constructors).

Expressions

exprDependencies :: Expr -> [QName]

Extracts the names of all variables, functions and constructors used in the given expression.

This also includes the names of predefined functions, error terms like undefined and error as well as constructors.

vars :: Expr -> [QName]

Extracts the names of all variables used in the given expression.

This also includes the names of functions, predefined functions and the error terms undefined and error.

varSet :: Expr -> Set QName

Like vars but returns a set of variable names.

cons :: Expr -> [QName]

Extracts the names of all constructors used in the given expression.

This also includes predefined constructors.

Type declarations

typeDeclDependencies :: TypeDecl -> [QName]

Extracts the dependencies of the given data type or type synonym declaration on type constructors (if a constructor contains undeclared type variables, the returned set contains the names of those type variables as well).

Returns an empty set if the given declaration is not a type declaration.

typeDeclDependencySet :: TypeDecl -> Set QName

Like typeDeclDependencies but returns the type's dependencies as a Set.

Function declarations

funcDeclDependencies :: FuncDecl -> [QName]

Extracts the dependencies of the given function declaration on constructors and other functions.

Returns an empty set if the given declaration is not a function declaration.

funcDeclDependencySet :: FuncDecl -> Set QName

Like funcDeclDependencies but returns the function's dependencies as a Set.

Modules

moduleDependencies :: Module -> [ModName]

Extracts the dependencies of the given module on other modules.

Every module depends on the Prelude module implicitly.

moduleDependencySet :: Module -> Set ModName

Like moduleDependencies but returnes a set of imported modules.