Compiler.Analysis.DependencyGraph

Description

This module contains functions to construct dependency graphs of Haskell declarations and modules. A dependency graph is a directed graph whose nodes are labeled with declarations. There is an edge from node A to B if the declaration of A depends on the declaration of B (i.e. in Coq B has to be defined before A or both have to be declared in the same sentence). The module dependency graph contains an edge from node A to B if the module of A contains an import declaration for the module of B.

The dependency graph does only contain defined top-level declarations (i.e. there are no nodes for build-in data types or operations and there are no nodes for local variables such as function parameters or variable patterns). However the entries of the dependency graph contain keys for predefined functions (but not local variables) and the special functions error and undefined that are used in error terms.

We distinguish between the type and function dependency graph. This is because in Haskell function declarations and type declarations live in separate scopes and we want to avoid name conflicts. Because we assume all type declarations to precede function declarations in the generated Coq code, this should not be a problem. For the same reason the function dependency graph does not include nodes for constructors (as always, the keys of used constructors are still present).

The construction of a dependency graph does not fail even if there are are undefined identifiers.

For debugging purposes dependency graphs can be converted to the DOT format such that they can be visualized using Graphviz (See https://www.graphviz.org/).

Documentation

type DGKey = QName

Every node of the dependency graph is uniquely identified by a key. We use the Haskell identifiers and symbols to identify the nodes.

type DGEntry node = (node, DGKey, [DGKey])

Every node (declaration) in a dependency graph is associated with a unique key (Haskell identifier) and a list of keys that identify the nodes this node depends on (adjacency list).

data DependencyGraph node

A dependency graph is a directed graph whose nodes are Haskell declarations (Usually TypeDecl or FuncDecl). There is an edge from node A to node B if the declaration of A depends on B.

Nodes are identified by their Haskell identifier (See DGKey). Internally nodes are identified by a number (See Vertex).

In addition to the actual Graph that stores the adjacency matrix of the internal identifiers, this tuple contains functions to convert between the internal and high level representation.

Constructors

DependencyGraph

Fields

Graph

The actual graph.

(Vertex -> DGEntry node)

Gets an entry for a vertex of the graph.

(DGKey -> Maybe Vertex)

Gets the vertex of a node with the given key.

Getters

dependencyGraphEntries :: DependencyGraph node -> [DGEntry node]

Gets the entries of the given dependency graph.

Predicates

dependsDirectlyOn

Arguments

:: DependencyGraph node
The dependency graph.

-> DGKey
The key of the first node.

-> DGKey
The key of the second node.

-> Bool

Tests whether two nodes (identified by the given keys) of the dependency graph depend on each other directly (i.e., there is an edge from the first node to the second node).

Returns False if one of the nodes does not exist.

Constructors

typeDependencyGraph :: [TypeDecl] -> DependencyGraph TypeDecl

Creates the dependency graph for a list of data type or type synonym declarations.

funcDependencyGraph :: [FuncDecl] -> DependencyGraph FuncDecl

Creates the dependency graph for a list of function declarations.

moduleDependencyGraph :: [Module] -> DependencyGraph Module

Creates the dependency graph for a list of Haskell modules.

Modules without a name are excluded from the dependency graph.