Compiler.Analysis.TypeInference

Description

This module contains functions for infering the type of expressions and function declarations as well as functions for annotating the types of variable patterns and adding visible applications for type arguments.

Function declarations

inferFuncDeclTypes :: [FuncDecl] -> Converter [TypeSchema]

Tries to infer the types of (mutually recursive) function declarations.

annotateFuncDeclTypes :: [FuncDecl] -> Converter [FuncDecl]

Infers the types of type arguments to functions and constructors used by the right-hand side of the given function declaration.

annotateFuncDeclTypes' :: [FuncDecl] -> Converter ([FuncDecl], [TypeSchema])

Like annotateFuncDeclTypes but also returns the type of the function declaration.

Expressions

inferExprType :: Expr -> Converter TypeSchema

Tries to infer the type of the given expression from the current context and abstracts it's type to a type schema.

The type of an expression e is the same type inferred by inferFuncDeclTypes for a function f defined by f = e where f does not occur free in e.

annotateExprTypes :: Expr -> Converter Expr

Infers the types of type arguments to functions and constructors used by the given expression.

Returns an expression where the type arguments of functions and constructors are applied explicitly.

annotateExprTypes' :: Expr -> Converter (Expr, TypeSchema)

Like annotateExprTypes but also returns the type of the expression.