Compiler.Backend.Coq.Converter.TypeSchema

Description

This module contains functions for converting between type expressions and type schemas.

Instantiating type schemas

instantiateTypeSchema :: TypeSchema -> Converter Type

Replaces the type variables in the given type schema by fresh type variables.

instantiateTypeSchema' :: TypeSchema -> Converter (Type, [Type])

Like instantiateTypeSchema but also returns the fresh type variables, the type schema has been instantiated with.

Abstracting type expressions

abstractTypeSchema :: [QName] -> Type -> Converter TypeSchema

Normalizes the names of type variables in the given type and returns it as a type schema.

The first argument contains the names of type variables that should be bound by the type schema. Usually these are the type variables that occur in the given type (see freeTypeVars).

Fresh type variables used by the given type are replaced by regular type variables with the prefix freshTypeArgPrefix. All other type variables are not renamed.

TODO move out of Converter monad.

abstractTypeSchema' :: [QName] -> Type -> Converter (TypeSchema, Subst Type)

Like abstractTypeSchema but returns the resulting type schema and the substitution that replaces the abstracted type variables by their name in the type schema.