Compiler.Backend.Coq.Keywords

Description

This module exports lists for keywords of the Gallina specification langauge and Vernacular commands.

Documentation

coqKeywords :: [String]

Keywords of the Gallina specification language.

This list has been taken from the Coq documentation https://coq.inria.fr/refman/language/gallina-specification-language.html.

vernacularCommands :: [String]

Identifiers that should not be used in Coq, because they are used in a command of The Vernacular.

This list contains all words of the commands listed in https://coq.inria.fr/refman/coq-cmdindex.html.