Compiler.Backend.Coq.Pretty

Description

This module contains a Pretty instance for nodes of the Coq AST.

Since we want to have a pretty instance for all Gallina instances but want to avoid overlapping instances, there is a wrapper type PrettyCoq. To pretty print a node x of the Coq AST writer pretty (PrettyCoq x).

Documentation

newtype PrettyCoq a

Wrapper data type that makes a Coq AST node of type a pretty printable.

Constructors

PrettyCoq

Fields

unwrapPrettyCoq :: a