Compiler.Converter

Description

This module exports functions for converting Haskell to Coq using the Free monad.

Modules

convertModule :: Module -> Converter [Sentence]

Converts a Haskell module to a Gallina sentences.

convertDecls :: [TypeDecl] -> [TypeSig] -> [FuncDecl] -> Converter [Sentence]

Converts the given declarations of a Haskell module.

Data type declarations

convertTypeDecls :: [TypeDecl] -> Converter [Sentence]

Converts the given data type or type synonym declarations.

convertTypeComponent :: DependencyComponent TypeDecl -> Converter [Sentence]

Converts a strongly connected component of the type dependency graph.

convertDataDecls :: [TypeDecl] -> Converter [Sentence]

Converts multiple (mutually recursive) Haskell data type declaration declarations.

Before the declarations are actually translated, their identifiers are inserted into the current environement. Otherwise the data types would not be able to depend on each other. The identifiers for the constructors are inserted into the current environmen as well. This makes the constructors more likely to keep their original name if there is a type variable with the same (lowercase) name.

After the Inductive sentences for the data type declarations there is one Arguments sentence and one smart constructor declaration for each constructor declaration of the given data types.

convertDataDecl :: TypeDecl -> Converter (IndBody, [Sentence])

Converts a Haskell data type declaration to the body of a Coq Inductive sentence, the Arguments sentences for it's constructors and the smart constructor declarations.

This function assumes, that the identifiers for the declared data type and it's (smart) constructors are defined already (see defineTypeDecl). Type variables declared by the data type or the smart constructors are not visible outside of this function.

Function declarations

convertFuncDecls :: [FuncDecl] -> Converter [Sentence]

Converts the given function declarations.

convertFuncComponent :: DependencyComponent FuncDecl -> Converter [Sentence]

Converts a strongly connected component of the function dependency graph.

convertNonRecFuncDecl :: FuncDecl -> Converter Sentence

Converts a non-recursive Haskell function declaration to a Coq Definition sentence.

convertRecFuncDecls :: [FuncDecl] -> Converter [Sentence]

Converts (mutually) recursive Haskell function declarations to Coq.

The function declarations are analysed first. If they contain constant arguments (i.e. arguments that are passed unchanged betwen recursive calls), they are converted using a Section sentence. Otherwise they are converted into helper and main functions.

Type expressions and type schemas

convertType :: Type -> Converter Term

Converts a Haskell type to Coq, lifting it into the Free monad.

\(\tau^\dagger = Free\,Shape\,Pos\,\tau^*\)

A type \(\tau\) is converted by lifting it into the Free monad and recursivly converting the argument and return types of functions using convertType'.

convertType' :: Type -> Converter Term

Converts a Haskell type to Coq.

In contrast to convertType, the type itself is not lifted into the Free moand. Only the argument and return types of contained function type constructors are lifted recursivly.

\(\alpha^* = \alpha'\)

A type variable \(\alpha\) is translated by looking up the corresponding Coq identifier \(\alpha'\).

\(T^* = T'\,Shape\,Pos\)

A type constructor \(T\) is translated by looking up the corresponding Coq identifier \(T'\) and adding the parameters \(Shape\) and \(Pos\).

\((\tau_1\,\tau_2)^* = \tau_1^*\,\tau_2^*\)

Type constructor applications are translated recursively but remain unchanged otherwise.

\((\tau_1 \rightarrow \tau_2)^* = \tau_1^\dagger \rightarrow \tau_2^\dagger\)

Type constructor applications are translated recursively but remain unchanged otherwise.

convertTypeSchema :: TypeSchema -> Converter Term

Converts a Haskell type schema to Coq.

Expressions

convertExpr :: Expr -> Converter Term

Converts a Haskell expression to Coq.