Compiler.Converter.Arg

Description

This module contains functions for converting function, constructor and type arguments from Haskell to Coq.

Documentation

convertTypeVarDecls

Arguments

:: Explicitness
Whether to generate an explicit or implit binder.

-> [TypeVarDecl]
The type variable declarations.

-> Converter [Binder]

Converts the declarations of type variables in the head of a data type or type synonym declaration to a Coq binder for a set of explicit or implicit type arguments.

E.g. the declaration of the type variable a in data D a = ... is translated to the binder (a : Type). If there are multiple type variable declarations as in data D a b = ... they are grouped into a single binder (a b : Type) because we assume all Haskell type variables to be of kind *.

The first argument controlls whether the generated binders are explicit (e.g. (a : Type)) or implicit (e.g. {a : Type}).

generateTypeVarDecls

Arguments

:: Explicitness
Whether to generate an explicit or implit binder.

-> [TypeVarIdent]
The names of the type variables to declare.

-> Converter [Binder]

Generates explicit or implicit Coq binders for the type variables with the given names that are either declared in the head of a data type or type synonym declaration or occur in the type signature of a function.

The first argument controlls whether the generated binders are explicit (e.g. (a : Type)) or implicit (e.g. {a : Type}).

generateTypeVarDecls'

Arguments

:: Explicitness
Whether to generate an explicit or implit binder.

-> [SrcSpan]
The location of the type variable declarations.

-> [TypeVarIdent]
The names of the type variables to declare.

-> Converter [Binder]

Like generateTypeVarDecls but also accepts the location of the type variables (for error reporting purposes).

convertArgs

Arguments

:: [VarPat]
The function arguments.

-> [Maybe Type]
The types of the function arguments.

-> Maybe Int
The position of the decreasing argument.

-> Converter [Binder]

Converts the arguments (variable patterns) of a potentially recursive function with the given types to explicit Coq binders.

If the type of an argument is not known, it's type will be inferred by Coq.

If the function is recursive, its decreasing argument (given index), is not lifted.

convertInferredArg :: VarPat -> Converter Binder

Converts the argument of a function (a variable pattern) to an explicit Coq binder whose type is inferred by Coq.

convertTypedArg :: VarPat -> Type -> Converter Binder

Converts the argument of a function (a variable pattern) with the given type to an explicit Coq binder.

convertDecArg :: VarPat -> Maybe Type -> Converter Binder

Convert the decreasing argument (variable pattern) if a recursive function with the given type to an explicit Coq binder.

In contrast to a regular typed argument (see convertTypedArg), the decreasing argument is not lifted to the Free monad. It is also registered as a non-monadic value.

convertArg :: VarPat -> Maybe Term -> Converter Binder

Converts the argument of a function (a variable pattern) to an explicit Coq binder.

convertPureArg :: VarPat -> Maybe Term -> Converter Binder

Like convertArg but marks the variable as non-monadic.

convertArg' :: Bool -> VarPat -> Maybe Term -> Converter Binder

Like convertArg but the first argument indicates whether the varibale has been lifted to the free monad or not (True for pure values and False for lifted values).

generateArgBinder :: Qualid -> Maybe Term -> Converter Binder

Generates an explicit Coq binder for a function argument with the given name and optional Coq type.

If no type is provided, it will be inferred by Coq.

convertAnonymousArg :: Maybe Type -> Converter (Qualid, Binder)

Converts the argument of an artifically generated function to an explicit Coq binder. A fresh Coq identifier is selected for the argument and returned together with the binder.