Compiler.Converter.Expr

Description

This module contains functions for converting Haskell expressions to Coq.

Documentation

etaConvert :: Expr -> Converter Expr

Applies eta-abstractions to a function or constructor application util the function or constructor is fully applied.

E.g. an application f x of a binary function f will be converted to y -> f x y where y is a fresh variable. This function does not apply the eta-conversion recursively.

No eta-conversions are applied to nested expressions.

convertExpr :: Expr -> Converter Term

Converts a Haskell expression to Coq.

convertExpr' :: Expr -> [Type] -> [Expr] -> Converter Term

Converts the application of a Haskell expression to the given arguments and visibly applied type arguments to Coq.

This function assumes the outer most expression to be fully applied by the given arguments (see also etaConvert).

generateApply :: Term -> [Term] -> Converter Term

Generates a Coq term for applying a monadic term to the given arguments.

generateApplyN :: Int -> Term -> [Term] -> Converter Term

Generates a Coq term for applying a function with the given arity to the given arguments.

If there are too many arguments, the remaining arguments are applied using generateApply. There should not be too few arguments (i.e. the function should be fully applied).

convertAlt :: Alt -> Converter Equation

Converts an alternative of a Haskell case-expressions to Coq.

convertConPat :: ConPat -> [VarPat] -> Converter Pattern

Converts a Haskell constructor pattern with the given variable pattern arguments to a Coq pattern.

convertVarPat :: VarPat -> Converter Pattern

Converts a Haskell variable pattern to a Coq variable pattern.

The types of variable patterns are not annotated in Coq.