Compiler.Converter.FuncDecl.Common

Description

This function contains auxilary functions that are used both to translate recursive and non-recursive Haskell functions to Coq.

Function environment entries

defineFuncDecl :: FuncDecl -> Converter ()

Inserts the given function declaration into the current environment.

Code generation

convertFuncHead :: FuncDecl -> Converter (Qualid, [Binder], Maybe Term)

Converts the name, arguments and return type of a function to Coq.

This code is shared between the conversion functions for recursive and no recursive functions (see convertNonRecFuncDecl and convertRecFuncDecls).