Compiler.Converter.Module

Description

This module contains functions for converting Haskell modules to Coq.

Documentation

convertModule :: Module -> Converter [Sentence]

Converts a Haskell module to a Gallina sentences.

addDecArgPragma :: [FuncDecl] -> Pragma -> Converter ()

Inserts the decreasing argument's index annotated by the given pragma into the environment.

Decreasing arguments can be annotated for all function declarations in the current module (first argument).

All other pragmas are ignored.

convertDecls :: [TypeDecl] -> [TypeSig] -> [FuncDecl] -> Converter [Sentence]

Converts the given declarations of a Haskell module.

convertTypeDecls :: [TypeDecl] -> Converter [Sentence]

Converts the given data type or type synonym declarations.

convertImportDecls :: [ImportDecl] -> Converter [Sentence]

Converts the given

convertImportDecl :: ImportDecl -> Converter Sentence

Convert a import declaration.

generateImport :: ModuleIdent -> ModName -> Converter Sentence

Generates a Coq import sentence for the module with the given name from the given library.

Modules from the base library are imported via From Base Require Import sentences and all other modules are also exported.

exportInterface :: Converter ModuleInterface

Generates the module interface exported by the currently translated module.

The interface contains all (non-internal) top-level entries that are currently in the environment. Only entries that are defined in the currently translated module are listed as exported. All other entries are "hidden". Hidden entries are included such that module interfaces are self contained and type synonyms can be expanded properly. All references in the entries are converted to fully qualified identifiers before they are exported.