Compiler.Converter.QuickCheck

Description

This module contains functions for converting simple QuickCheck properties (i.e. Haskell functions that start with prop_) to Coq Theorem sentences.

Documentation

quickCheckModuleName :: ModName

The name of the QuickCheck module.

quickCheckInterface :: ModuleInterface

Environment for the Test.QuickCheck module.

Only the Property data type is exported. There are no entries for the supported QuickCheck operators, they are handled directly by the converter.

propertyName :: Name

The name of the Property data type.

propertyQName :: QName

The qualified name of the Property data type.

isQuickCheckProperty :: FuncDecl -> Converter Bool

Tests whether the given declaration is a QuickCheck property.

QuickCheck properties are functions with the prefix prop_ or which return a value of type Property.

containsQuickCheckProperty :: DependencyComponent FuncDecl -> Converter Bool

Tests whether the given strongly connected component of the function dependency graph contains a QuickCheck property.

filterQuickCheckProperties :: [DependencyComponent FuncDecl] -> Converter ([FuncDecl], [DependencyComponent FuncDecl])

Partitions the given list of strongly connected components of the function dependency graph into a list of QuickCheck properties and dependency components that contain no QuickCheck properties.

Reports a fatal error message if there is there is a (mutually) recursive QuickCheck property in one of the components.

convertQuickCheckProperty :: FuncDecl -> Converter [Sentence]

Converts the given QuickCheck property to a Coq Theorem with an empty Proof (or the proof configured in the `.proofs.toml` file).

convertQuickCheckExpr :: Expr -> Converter Term

Converts the expression on the right hand side of a QuickCheck property declaration.

convertQuickCheckBool :: Expr -> Converter Term

Converts a boolean expression to a Coq property that states that the expression must evaluate to true.

convertQuickCheckPrecond :: Expr -> Expr -> Converter Term

Converts a QuickCheck precondition to a Coq implication.

convertQuickCheckEq :: Expr -> Expr -> Converter Term

Converts the QuickCheck equality (===) to a Coq reflexive equality.

convertQuickCheckNeq :: Expr -> Expr -> Converter Term

Converts the QuickCheck equality (=/=) to a Coq reflexive inequality.

convertQuickCheckConj :: Expr -> Expr -> Converter Term

Converts the QuickCheck conjunction (.&&.) to a Coq conjunction.

convertQuickCheckDisj :: Expr -> Expr -> Converter Term

Converts the QuickCheck conjunction (.||.) to a Coq conjunction.