Compiler.Coq.AST

Description

This module contains smart constructors for nodes of the Coq AST. For convenience the original Coq AST is exported as well.

Comments

comment :: String -> Sentence

Smart constructor for Coq comments.

Proofs.

blankProof :: Proof

An admitted proof that contains only a placeholder text.

Identifiers

ident :: String -> Ident

Smart constructor for unqualified Coq identifiers.

bare :: String -> Qualid

Smart constructor for Coq identifiers.

unpackQualid :: Qualid -> Maybe String

Gets the identifier for the given unqualified Coq identifier. Returns Nothing if the given identifier is qualified.

Functions

app :: Term -> [Term] -> Term

Smart constructor for the application of a Coq function or (type) constructor.

If the first argument is an application term, the arguments are added to that term. Otherwise a new application term is created.

arrows

Arguments

:: [Term]
The types of the function arguments.

-> Term
The return type of the function.

-> Term

Smart constructor for a Coq function type.

fun :: [Qualid] -> [Maybe Term] -> Term -> Term

Smart constructor for the construction of a Coq lambda expression with the given arguments and right hand side.

The second argument contains the types of the arguments are inferred by Coq.

inferredFun :: [Qualid] -> Term -> Term

Like fun, but all argument types are inferred.

Binders

typedBinder :: Explicitness -> [Qualid] -> Term -> Binder

Smart constructor for an explicit or implicit typed Coq binder.

typedBinder' :: Explicitness -> Qualid -> Term -> Binder

Like typedBinder but for a single identifier.

Assumptions

variable :: [Qualid] -> Term -> Sentence

Generates a Variable assumption sentence.

Definition sentences

definitionSentence

Arguments

:: Qualid
The name of the definition.

-> [Binder]
Binders for the parameters of the definition.

-> Maybe Term
The return type of the definition.

-> Term
The right hand side of the definition.

-> Sentence

Smart constructor for a Coq definition sentence.

Types

sortType :: Term

The type of a type variable.

Expressions

string :: String -> Term

Smart constructor for Coq string literals.

equation :: Pattern -> Term -> Equation

Smart constructor for Coq equations for match expressions.

match :: Term -> [Equation] -> Term

Smart constructor for a Coq match expression.

equals :: Term -> Term -> Term

Smart constructor for reflexive equality in Coq.

notEquals :: Term -> Term -> Term

Smart constructor for reflexive inequality in Coq.

conj :: Term -> Term -> Term

Smart constructor for a conjunction in Coq.

disj :: Term -> Term -> Term

Smart constructor for a disjunction in Coq.

Imports

requireImport :: [ModuleIdent] -> Sentence

Creates a Require Export ... sentence.

This does not generate a Require Import ... sentence since modules depending on the importing module could use definitions from the imported modules.

requireImportFrom :: ModuleIdent -> [ModuleIdent] -> Sentence

Creates a From ... Require Import ... sentence.

Since From ... Require ... sentences are only generated for built in modules, it does not have to Export the required module.