Compiler.Environment

Description

This module contains a data type that encapsulates the state of the compiler. There are also utility functions to modify the state and retreive information stored in the state.

Module interface

data ModuleInterface

Data type that contains the information of a module environment that is exported and imported.

Constructors

ModuleInterface

Fields

interfaceModName :: ModName

The name of the module.

interfaceExports :: Set ScopedName

The names (qualified with their original module name) that are exported by the module.

interfaceEntries :: Set EnvEntry

The entries (including hidden entries) defined in or imported by the module.

Environment

data Environment

Data type that encapsulates the state of the converter.

Constructors

Environment

Fields

envDepth :: Int

The number of parent environments.

envModName :: ModName

The name of the currently translated module. Defaults to the empty string.

envAvailableModules :: Map ModName ModuleInterface

Maps names of modules that can be imported to their interface.

envInSection :: Bool

Whether the currently converted node is inside of a Section sentence.

envEntries :: Map ScopedName (Set EnvEntry, Int)

Maps Haskell names to entries for declarations. In addition to the entry, the envDepth of the environment is recorded. There can be multiple entries with the same name as long as they are not referenced. Entries are identified by their original name.

envTypeSigs :: Map QName TypeSchema

Maps names of Haskell functions to their annotated types.

envDecArgs :: Map QName (Int, String)

Maps Haskell function names to the index and name of their decreasing argument. Contains no entry for non-recursive functions, but there are also entries for functions that are shadowed by local variables.

envFreshIdentCount :: Map String Int

The number of fresh identifiers that were used in the environment with a certain prefix.

envQuickCheckEnabled :: Bool

Whether the translation of QuickCheck properties is enabled in the current environment (i.e. the module imports Test.QuickCheck).

emptyEnv :: Environment

An environment that does not even contain any predefined types and functions.

childEnv :: Environment -> Environment

Creates a child environment of the given environment.

Module information

makeModuleAvailable :: ModuleInterface -> Environment -> Environment

Inserts the interface of a module name into the environment such that it can be imported.

isModuleAvailable :: ModName -> Environment -> Bool

Tests whether the module with the given name can be imported.

lookupAvailableModule :: ModName -> Environment -> Maybe ModuleInterface

Looks up the environment of another module that can be imported.

Inserting entries into the environment

addEntry :: QName -> EnvEntry -> Environment -> Environment

Inserts an entry into the given environment and associates it with the given name.

addEntry' :: QName -> EnvEntry -> Int -> Environment -> Environment

Like addEntry but has an additional parameter for the envDepth value to record.

defineDecArg :: QName -> Int -> String -> Environment -> Environment

Stores the index of the decreasing argument of a recursive function in the environment.

removeDecArg :: QName -> Environment -> Environment

Removes the index of the decreasing argument of a recursive function from the environment (e.g., because the function has been transformed into a non-recursive function).

defineTypeSig :: QName -> TypeSchema -> Environment -> Environment

Inserts the given type signature into the environment.

Looking up entries from the environment

lookupEntries :: Scope -> QName -> Environment -> [EnvEntry]

Looks up the entries that have been associated with the given name in the specified scope of the given environment.

lookupEntry :: Scope -> QName -> Environment -> Maybe EnvEntry

existsLocalEntry :: Scope -> QName -> Environment -> Bool

Tests whether there is an entry with the given name in the current environment that was not inherited from a parent environment.

refersTo :: QName -> Scope -> QName -> Environment -> Bool

Tests whether the given name (third argument) refers to an entry with the given original name (first argument).

isFunction :: QName -> Environment -> Bool

Tests whether the given name identifies a function in the given environment.

Returns False if there is no such function.

isVariable :: QName -> Environment -> Bool

Tests whether the given name identifies a local variable in the given environment.

Returns False if there is no such variable.

hasTypeSig :: QName -> Environment -> Bool

Tests whether there is a type signature for the given identifier.

isPureVar :: QName -> Environment -> Bool

Test whether the variable with the given name is not monadic.

lookupIdent :: Scope -> QName -> Environment -> Maybe Qualid

Looks up the Coq identifier for a Haskell function, (type) constructor or (type) variable with the given name.

Returns Nothing if there is no such function, (type/smart) constructor, constructor or (type) variable with the given name.

lookupSmartIdent :: QName -> Environment -> Maybe Qualid

Looks up the Coq identifier for the smart constructor of the Haskell constructor with the given name.

Returns Nothing if there is no such constructor.

usedIdents :: Environment -> [Qualid]

Gets a list of Coq identifiers for functions, (type/smart) constructors, (type/fresh) variables that were used in the given environment already.

lookupSrcSpan :: Scope -> QName -> Environment -> Maybe SrcSpan

Looks up the location of the declaration with the given name.

lookupTypeArgs :: Scope -> QName -> Environment -> Maybe [TypeVarIdent]

Looks up the type variables used by the type synonym, (smart) constructor or type signature of the function with the given name.

Returns Nothing if there is no such type synonym, function or (smart) constructor with the given name.

lookupTypeArgArity :: Scope -> QName -> Environment -> Maybe Int

Returns the length of the list returned by lookupTypeArgs.

lookupArgTypes :: Scope -> QName -> Environment -> Maybe [Maybe Type]

Looks up the argument and return types of the function or (smart) constructor with the given name.

Returns Nothing if there is no such function or (smart) constructor with the given name.

lookupReturnType :: Scope -> QName -> Environment -> Maybe Type

Looks up the return type of the function or (smart) constructor with the given name.

Returns Nothing if there is no such function or (smart) constructor with the given name or the return type is not known.

lookupTypeSchema :: Scope -> QName -> Environment -> Maybe TypeSchema

Gets the type schema of the variable, function or constructor with the given name.

lookupArity :: Scope -> QName -> Environment -> Maybe Int

Looks up the number of arguments expected by the Haskell function or smart constructor with the given name.

Returns Nothing if there is no such function or (smart) constructor with the given name.

lookupTypeSynonym :: QName -> Environment -> Maybe ([TypeVarIdent], Type)

Looks up the type the type synonym with the given name is associated with.

Returns Nothing if there is no such type synonym.

lookupTypeSig :: QName -> Environment -> Maybe TypeSchema

Looks up the annotated type of a user defined Haskell function with the given name.

Returns Nothing, if there is no such type signature or the entry has been replaced already.

needsFreeArgs :: QName -> Environment -> Bool

Tests whether the function with the given name needs the arguments of the Free monad.

Returns False if there is no such function.

isPartial :: QName -> Environment -> Bool

Tests whether the function with the given name is partial.

Returns False if there is no such function.

lookupDecArg :: QName -> Environment -> Maybe (Int, String)

Looks up the index and name of the decreasing argument of the recursive function with the given name.

Returns Nothing if there is no such recursive function.

lookupDecArgIndex :: QName -> Environment -> Maybe Int

Like lookupDecArg but returns the decreasing argument's index only.

lookupDecArgIdent :: QName -> Environment -> Maybe String

Like lookupDecArg but returns the decreasing argument's name only.

QuickCheck support

enableQuickCheck :: Environment -> Environment

Enables the translation of QuickCheck properties.

isQuickCheckEnabled :: Environment -> Bool

Tests whether the translation of QuickCheck properties is enabled in the given environment.

This flag is usually set to True if there is a import Test.QuickCheck declaration.