Compiler.Environment.Entry

Description

This module contains data types that are used to store information about declared functions, (type) variables and (type) constructors in the environment.

Documentation

data EnvEntry

Entry of the environment.

Constructors

DataEntry
Entry for a data type declaration.

Fields

entrySrcSpan :: SrcSpan

The source code location where the data type was declared.

entryArity :: Int

The number of type arguments expected by the type constructor.

entryIdent :: Qualid

The name of the data type in Coq.

entryName :: QName

The name of the data type in the module it has been defined in.

TypeSynEntry
Entry for a type synonym declaration.

Fields

entrySrcSpan :: SrcSpan

The source code location where the data type was declared.

entryArity :: Int

The number of type arguments expected by the type constructor.

entryTypeArgs :: [TypeVarIdent]

The names of the type arguments.

entryTypeSyn :: Type

The type that is abbreviated by this type synonym.

entryIdent :: Qualid

The name of the data type in Coq.

entryName :: QName

The name of the data type in the module it has been defined in.

TypeVarEntry
Entry for a type variable.

Fields

entrySrcSpan :: SrcSpan

The source code location where the data type was declared.

entryIdent :: Qualid

The name of the data type in Coq.

entryName :: QName

The name of the data type in the module it has been defined in.

ConEntry
Entry for a data constructor.

Fields

entrySrcSpan :: SrcSpan

The source code location where the data type was declared.

entryArity :: Int

The number of type arguments expected by the type constructor.

entryArgTypes :: [Maybe Type]

The types of the constructor's arguments (if known). Contains exactly entryArity elements.

entryReturnType :: Maybe Type

The return type of the data constructor (if known).

entryIdent :: Qualid

The name of the data type in Coq.

entrySmartIdent :: Qualid

The name of the corresponding smart constructor in Coq.

entryName :: QName

The name of the data type in the module it has been defined in.

FuncEntry
Entry for a function declaration.

Fields

entrySrcSpan :: SrcSpan

The source code location where the data type was declared.

entryArity :: Int

The number of type arguments expected by the type constructor.

entryTypeArgs :: [TypeVarIdent]

The names of the type arguments.

entryArgTypes :: [Maybe Type]

The types of the constructor's arguments (if known). Contains exactly entryArity elements.

entryReturnType :: Maybe Type

The return type of the data constructor (if known).

entryNeedsFreeArgs :: Bool

Whether the arguments of the Free monad need to be passed to the function.

entryIsPartial :: Bool

Whether the function is partial, i.e., requires an instance of the Partial type class when translated to Coq.

entryIdent :: Qualid

The name of the data type in Coq.

entryName :: QName

The name of the data type in the module it has been defined in.

VarEntry
Entry for a variable.

Fields

entrySrcSpan :: SrcSpan

The source code location where the data type was declared.

entryIsPure :: Bool

Whether the variable has not been lifted to the free monad.

entryIdent :: Qualid

The name of the data type in Coq.

entryName :: QName

The name of the data type in the module it has been defined in.

entryScope :: EnvEntry -> Scope

Gets the scope an entry needs to be defined in.

entryScopedName :: EnvEntry -> ScopedName

Gets the scope and name of the given entry.

isDataEntry :: EnvEntry -> Bool

Tests whether the given entry of the environment describes a data type.

isTypeSynEntry :: EnvEntry -> Bool

Tests whether the given entry of the environment describes a type synonym.

isTypeVarEntry :: EnvEntry -> Bool

Tests whether the given entry of the environment describes a type variable.

isConEntry :: EnvEntry -> Bool

Tests whether the given entry of the environment describes a data constructor.

isFuncEntry :: EnvEntry -> Bool

Tests whether the given entry of the environment describes a function.

isVarEntry :: EnvEntry -> Bool

Tests whether the given entry of the environment describes a variable.

isTopLevelEntry :: EnvEntry -> Bool

Tests whether the given entry of the environment describes a top-level data type, type synonym, constructor or function.

Type variables and local variables are no top level entries.

entryHasSmartIdent :: EnvEntry -> Bool

Tests whether the given entry distinguishes between a entryIdent and entrySmartIdent.

prettyEntryType :: EnvEntry -> String

Gets a human readable description of the entry type.