Compiler.Environment.Fresh

Description

This module contains functions for generating frsh identifiers.

Fresh identifiers are identifiers that have been introduced aritifically into the Haskell or Coq AST. They are guranteed not to conflict with any other valid identifier.

Documentation

freshArgPrefix :: String

The prefix to use for artificially introduced variables of type a.

freshFuncPrefix :: String

The prefix to use for artificially introduced variables of type a -> b.

freshBoolPrefix :: String

The prefix to use for artificially introduced variables of type Bool.

freshHaskellIdent :: String -> Converter String

Gets the next fresh Haskell identifier from the current environment.

All fresh identifiers contain an at-sign (See internalIdentChar). This ensures that they cannot be confused with actual Haskell identifiers. The corresponding Coq identifier contains an underscore instead (see Compiler.Environment.Renamer). The at-sign (or underscore) is preceded by the given prefix and followed by an incrementing number (staring at 0).

If the given prefix is a fresh identifier itself (i.e. contains an at-sign), the prefix of that identifier is used instead.

The freshly generated identifier is not inserted into the environment, i.e. it can still be used to create a declaration.

freshCoqIdent :: String -> Converter String

Gets the next fresh Haskell identifier from the current environment and renames it such that it can be used in Coq.

The freshly generated identifier is not inserted into the environment, i.e. it can still be used to create a declaration.

freshCoqQualid :: String -> Converter Qualid

Like freshCoqIdent but the resulting Coq identifier is wrapped in a G.Qualid.