Compiler.Environment.Resolver

Description

This module contains functions for converting unqualified identifiers to qualified identifiers.

Resolving to original name

resolveReferences :: EnvEntry -> Converter EnvEntry

Resolves all type names referenced by the given entry of the environment to their original name.

resolveTypes :: Type -> Converter Type

Resolves all type names referenced by the given type expression to their original name.

Resolving by custom function

resolveReferencesWith :: (SrcSpan -> QName -> QName) -> EnvEntry -> EnvEntry

Resolves all type names referenced by the given entry of the environment by applying the given function.

All referenced entries must be in the environment and unambigious.

resolveReferencesWithM :: Monad m => (SrcSpan -> QName -> m QName) -> EnvEntry -> m EnvEntry

Resolves all type names referenced by the given entry of the environment by applying the given function monadically.

All referenced entries must be in the environment and unambigious.