Compiler.Frontend.Haskell.Pretty

Description

This module contains a Pretty instance for module nodes of the Haskell Source Extensions AST.

The pretty instance is used to dump the output of the pattern matching compiler library (See Compiler.Frontent.Haskell.PatternMatching).

Orphan instances

Pretty (Module l)
Pretty instance for module nodes of Haskell Source Extensions AST.

Instance details

Methods

pretty :: Module l -> Doc

prettyList :: [Module l] -> Doc