Compiler.Frontend.Haskell.SrcSpanConverter

Description

This module contains a type class for converting source spans from the haskell-src-exts package to the source spans of the intermediate representation of the compiler.

Documentation

class ConvertableSrcSpan ss where

Type class for haskell-src-exts source spans that can be converted to SrcSpans for pretty printing of messages.

Methods

convertSrcSpan

Arguments

:: [(String, [String])]
A map of file names to lines of source code.

-> ss
The original source span to convert.

-> SrcSpan

Converts a haskell-src-exts source span to a SrcSpan by attaching the corresponding line of source code.