Compiler.Haskell.AST

Description

An alternative AST that represents the subset of Haskell modules that comply with our restrictions on the accepted input format.

The aim of this module is to make the conversion functions easier to comprehend and reduce coupling with the haskell-src-ext package.

This AST cannot be parsed directly. Instances of the AST are usually created by converting an existing AST (e.g. from the haskell-src-ext package). For testing purposes instances may be created directly.

Documentation

data Name

An identifier or a symbolic name.

The constructors of this type do not contain source spans because Names are intended to be comparable. They are used as keys to identify nodes of the dependency graph for example.

Constructors

Ident String
An identifier, e.g. Ident "f" for a function f.

Symbol String
A symbolic name, e.g. Symbol "+" for (+).

identFromName :: Name -> Maybe String

Extracts an identifier from a Haskell name. Returns Nothing if the given name is a symbol and not an identifier.

data QName

A potentially qualified Name.

Constructors

Qual ModName Name
A qualified Name.

UnQual Name
An unqualified Name.

identFromQName :: QName -> Maybe String

Extracts an identifier from an unqualified Haskell name.

Returns Nothing if the given name is qualified or a symbol and not an identifier.

type TypeVarIdent = String

The name of a type variable.

type ModName = String

The name of a module.

type VarName = QName

The name of a function or build-in operator used in prefix notation, e.g. f x y or (+) n m

type ConName = QName

The name of an constructor used in prefix notation, e.g. (:) x xs.

type TypeConName = QName

The name of a type or type constructor, e.g. Integer or [] a

data DeclIdent

The name of a function, data type, type synonym or constructor defined by the user including location information.

Because the user cannot declare symbols at the moment, the constructor of this data type takes a String and not a Name.

Constructors

DeclIdent

Fields

declIdentSrcSpan :: SrcSpan

fromDeclIdent :: String

internalIdentChar :: Char

The character that is used to mark internal identifiers.

This is used for example to generate fresh identifiers that don't conflict with actual Haskell identifiers or to hide entries in module interfaces.

isInternalIdent :: String -> Bool

Tests whether the given Haskell identifier was generated for internal use only (i.e., contains internalIdentChar).

isInternalName :: Name -> Bool

Tests whether the given Haskell name was generated for interal use only (i.e., it is an identifier that matches isInternalIdent).

isInternalQName :: QName -> Bool

Tests whether the given qualified Hasell name was generted for internal use only (i.e., the name or module name are internal according to isInternalIdent and isInternalName, respectively).

data Module

A module declaration.

If the module has no module header, the module name defauts to Main.

Constructors

Module

Fields

modSrcSpan :: SrcSpan

modName :: ModName

modImports :: [ImportDecl]

modTypeDecls :: [TypeDecl]

modTypeSigs :: [TypeSig]

modPragmas :: [Pragma]

modFuncDecls :: [FuncDecl]

data Comment

A comment.

Comments are collected during parsing. However, the final AST contains no comments. Pragmas (see DecArgPragma) are extracted from comments during simplification.

Constructors

BlockComment
A multi-line comment (i.e., {- ... -}).

Fields

commentSrcSpan :: SrcSpan

commentText :: String

LineComment
A single-line comment (i.e., -- ...).

Fields

commentSrcSpan :: SrcSpan

commentText :: String

customPragmaPrefix :: String

All custom pragmas of the compiler start with HASKELL_TO_COQ.

data Pragma

Data type for custom {--} pragmas.

Constructors

DecArgPragma
A {--} or {--} pragma.

Fields

pragmaSrcSpan :: SrcSpan

decArgPragmaFuncName :: String

decArgPragmaArg :: Either String Int

prettyPragma :: Doc -> Doc

Pretty prints a custom {--} pragma with the given contents.

data ImportDecl

An import declaration.

Constructors

ImportDecl

Fields

importSrcSpan :: SrcSpan

importName :: ModName

type TypeVarDecl = DeclIdent

The name of a type variable declaration in the head of a data type or type synonym declaration including location information.

typeVarDeclToType :: TypeVarDecl -> Type

Converts the declaration of a type variable to a type.

data TypeDecl

A data type or type synonym declaration.

While it is allowed to define constructors in infix notation, data type and type synonym declarations must not be in infix notation. This is because the TypeOperators language extension is not supported.

Constructors

DataDecl

Fields

typeDeclSrcSpan :: SrcSpan

typeDeclIdent :: DeclIdent

typeDeclArgs :: [TypeVarDecl]

dataDeclCons :: [ConDecl]

TypeSynDecl

Fields

typeDeclSrcSpan :: SrcSpan

typeDeclIdent :: DeclIdent

typeDeclArgs :: [TypeVarDecl]

typeSynDeclRhs :: Type

data ConDecl

A constructor declaration.

Even though there is not a dedicated constructor, the constructor is allowed to be in infix notation, but the name of the constructor must not be a symbol.

Constructors

ConDecl

Fields

conDeclSrcSpan :: SrcSpan

conDeclIdent :: DeclIdent

conDeclFields :: [Type]

data TypeSig

A type signature of one or more function declarations.

Constructors

TypeSig

Fields

typeSigSrcSpan :: SrcSpan

typeSigDeclIdents :: [DeclIdent]

typeSigTypeSchema :: TypeSchema

data FuncDecl

A function declaration.

Even though there is not a separate constructor, it is allowed to define functions in infix notation e.g.:

    append :: [a] -> [a] -> [a]
    xs `append` ys = ...
  

But it is currently not supported to use symbolic names for such definitions. Thus the following is not allowed:

    (++) :: [a] -> [a] -> [a]
    xs ++ ys = ...
  

The precedence and associativity (fixity) of infix declarations can be specified as well.

    infixr 5 `append`
  

Constructors

FuncDecl

Fields

funcDeclSrcSpan :: SrcSpan

A source span that spans the entire function declaration.

funcDeclIdent :: DeclIdent

The name of the function.

funcDeclTypeArgs :: [TypeVarDecl]

The type arguments of the function.

funcDeclArgs :: [VarPat]

The function's argument patterns.

funcDeclRhs :: Expr

The right-hand side of the function declaration.

funcDeclReturnType :: Maybe Type

The return type of the function.

funcDeclName :: FuncDecl -> Name

Gets the name of the given function declaration.

funcDeclQName :: FuncDecl -> QName

Gets the unqualified name of the given function declaration.

funcDeclTypeSchema :: FuncDecl -> Maybe TypeSchema

Gets the type schema of the given function declaration or Nothing if at least one of the argument or the return type is not annoated.

data TypeSchema

A Haskell type expression with explicitly introduced type variables.

Constructors

TypeSchema

Fields

typeSchemaSrcSpan :: SrcSpan

typeSchemaArgs :: [TypeVarDecl]

typeSchemaType :: Type

data Type

A Haskell type expression.

Build-in types are represented by applications of their type constructors. E.g. the type [a] is represented as TypeApp (TypeCon "[]") (TypeVar "a"). The only exception to this rule is the function type a -> b. It is represented directly as FuncType (TypeVar "a") (TypeVar "b"). The syntax (->) a b is not supported at the moment. This is due to the special role of functions during the translation to Coq.

Constructors

TypeVar
A type variable.

Fields

typeSrcSpan :: SrcSpan

typeVarIdent :: TypeVarIdent

TypeCon
A type constructor.

Fields

typeSrcSpan :: SrcSpan

typeConName :: TypeConName

TypeApp
A type constructor application.

Fields

typeSrcSpan :: SrcSpan

typeAppLhs :: Type

typeAppRhs :: Type

FuncType
A function type.

Fields

typeSrcSpan :: SrcSpan

funcTypeArg :: Type

funcTypeRes :: Type

typeApp

Arguments

:: SrcSpan
The source span to insert into generated nodes.

-> Type
The partially applied type constructor.

-> [Type]
The type arguments to pass to the type constructor.

-> Type

Creates a type constructor application type.

The given source span is inserted into the generated type constructor and every generated type constructor application.

typeConApp

Arguments

:: SrcSpan
The source span to insert into generated nodes.

-> TypeConName
The name of the type constructor to apply.

-> [Type]
The type arguments to pass to the type constructor.

-> Type

Creates a type constructor application type for the constructor with the given name.

The given source span is inserted into the generated type constructor and every generated type constructor application.

funcType :: SrcSpan -> [Type] -> Type -> Type

Creates a function type with the given argument and return types.

splitFuncType :: Type -> Int -> ([Type], Type)

Splits the type of a function or constructor with the given arity into the argument and return types.

This is basically the inverse of funcType.

prettyTypePred :: Int -> Type -> Doc

Pretty prints a type and adds parenthesis if necessary.

The first argument indicates the precedence of the sourrounding context. * 0 - Top level. No parenthesis are neccessary. * 1 - Parenthesis are needed arround function types. * 2 - Parenthesis are also needed arround type constructor applications.

data Expr

A Haskell expression.

Even though there are no dedicated constructors, the infix applications of functions and constructors (including left and right sections) are supported. This kind of syntactic sugar is removed during simplification (see Compiler.Haskell.Simplifier).

Constructors

Con
A constructor.

Fields

exprSrcSpan :: SrcSpan

exprConName :: ConName

Var
A function or local variable.

Fields

exprSrcSpan :: SrcSpan

exprVarName :: VarName

App
Function or constructor application.

Fields

exprSrcSpan :: SrcSpan

exprAppLhr :: Expr

exprAppRhs :: Expr

TypeAppExpr
Visible type application.

Fields

exprSrcSpan :: SrcSpan

exprTypeAppLhs :: Expr

exprTypeAppRhs :: Type

If
if expression.

Fields

exprSrcSpan :: SrcSpan

ifExprCond :: Expr

ifExprThen :: Expr

ifExprElse :: Expr

Case
case expression.

Fields

exprSrcSpan :: SrcSpan

caseExprScrutinee :: Expr

caseExprAlts :: [Alt]

Undefined
Error term undefined.

Fields

exprSrcSpan :: SrcSpan

ErrorExpr
Error term error "message".

Fields

exprSrcSpan :: SrcSpan

errorExprMsg :: String

IntLiteral
An integer literal.

Fields

exprSrcSpan :: SrcSpan

intLiteralValue :: Integer

Lambda
A lambda abstraction.

Fields

exprSrcSpan :: SrcSpan

lambdaExprArgs :: [VarPat]

lambdaEprRhs :: Expr

ExprTypeSig
A type annotation.

Fields

exprSrcSpan :: SrcSpan

exprTypeSigExpr :: Expr

exprTypeSigTypeSchema :: TypeSchema

app :: SrcSpan -> Expr -> [Expr] -> Expr

Creates an expression for applying the given expression to the provided arguments.

The given source span is inserted into the generated function application expressions.

varApp

Arguments

:: SrcSpan
The source span to insert into generated nodes.

-> VarName
The name of the function to apply.

-> [Expr]
The arguments to pass to the function.

-> Expr

Creates an expression for applying the function with the given name.

The given source span is inserted into the generated function reference and every generated function application.

conApp

Arguments

:: SrcSpan
The source span to insert into generated nodes.

-> ConName
The name of the constructor to apply.

-> [Expr]
The arguments to pass to the constructor.

-> Expr

Creates a data constructor application expression.

The given source span is inserted into the generated constructor reference and every generated constructor application.

visibleTypeApp :: SrcSpan -> Expr -> [Type] -> Expr

Creates an expression for passing the type arguments of a function or constructor explicitly.

The given source span is inserted into every generated visible type application node.

prettyExprPred :: Int -> Expr -> Doc

Pretty prints an expression and adds parenthesis if necessary.

The first argument indicates the precedence of the sourrounding context. * 0 - Top level. No parenthesis are neccessary. * 1 - Parenthesis are needed around if, case and lambda expressions. * 2 - Parenthesis are also needed around function applications.

data Alt

One alternative of a case expression.

Every alternative of a case expression matches a constructor of the matched expression's data type. All arguments of the constructor pattern are variable patterns.

Constructors

Alt

Fields

altSrcSpan :: SrcSpan

altConPat :: ConPat

altVarPats :: [VarPat]

altRhs :: Expr

data ConPat

A constructor pattern used in an alternative of a case expression.

The main purpose of this data type is to add location information to a ConName.

Constructors

ConPat

Fields

conPatSrcSpan :: SrcSpan

conPatName :: ConName

conPatToExpr :: ConPat -> Expr

Converts a constructor pattern to a constructor expression.

data VarPat

A variable pattern used as an argument to a function, lambda abstraction or constructor pattern.

The variable pattern can optionally have a type signature.

Constructors

VarPat

Fields

varPatSrcSpan :: SrcSpan

varPatIdent :: String

varPatType :: Maybe Type

varPatName :: VarPat -> Name

Gets the name of the given variable pattern.

varPatQName :: VarPat -> QName

Gets the qualified name of the given variable pattern.

varPatToExpr :: VarPat -> Expr

Converts a variable pattern to a variable expression.

toVarPat :: String -> VarPat

Converts the given identifier to a variable pattern without type annotation.

preludeModuleName :: ModName

The name of the Prelude module.

TODO once import ... as ... is supported, the Prelude could be renamed by the user.

unitTypeConName :: TypeConName

The name of the unit type constructor.

tupleTypeConName :: Int -> TypeConName

The name of the n-ary tuple type constructor.

listTypeConName :: TypeConName

The name of the list type constructor.

unitConName :: ConName

Name of the unit data constructor.

nilConName :: ConName

The name of the empty list data constructor.

consConName :: ConName

The name of the non empty list data constructor.

tupleConName :: Int -> ConName

The name of the n-ary tuple data constructor.

integerTypeConName :: TypeConName

When inferring the type of IntegerLiterals this is the type to infer.

boolTypeConName :: TypeConName

When translating if expressions, we annotate the type of the condition with Bool.

negateOpName :: VarName

The unary prefix operator - is translated to the application of the negate function.

errorFuncName :: VarName

The name of the error function.

undefinedFuncName :: VarName

The name of the undefined function.