Compiler.Haskell.Inliner

Description

This module contains functions for inlining the definition of functions into expressions or other function declarations and for expanding type synonyms in type expressions, type signatures as well as data type and other type synonym declarations.

Inlining is performed during the translation of recursive function declarations to inline the definition of the non-recursive main function into the recursive helper functions.

The expansion of type synonyms is used to split the type of a function into the types of it's arguments and the return type and during the conversion of mutually recursive data type and type synonym declarations to break the dependency between of the data type declarations on the type synonyms (because they cannot be declared in the same sentence).

Documentation

inlineFuncDecls :: [FuncDecl] -> FuncDecl -> Converter FuncDecl

Inlines the right hand sides of the given function declarations into the right hand sides of other function declarations.

inlineExpr :: [FuncDecl] -> Expr -> Converter Expr

Inlines the right hand sides of the given function declarations into an expression.

expandAllTypeSynonymsInDecl :: TypeDecl -> Converter TypeDecl

Expands all type synonyms in all types in the given data type or type synonym declaration.

expandAllTypeSynonymsInConDecl :: ConDecl -> Converter ConDecl

Expands all type synonyms in all types in the given constructor declaration.

expandTypeSynonym :: Type -> Converter Type

Expands the outermost type synonym in the given type expression.

expandAllTypeSynonyms :: Type -> Converter Type

Expands all type synonyms used in the given type expression by the type they are associated with in the current environment.

expandTypeSynonyms :: Int -> Type -> Converter Type

Like expandAllTypeSynonyms but accepts an additional argument for the maximum depth.

If the maximum depth if 0, the type will be returned unchanged. If it is 1 only the outermost type synonym will be expanded. If it is negative, all type synonyms will be expanded (see also expandAllTypeSynonyms).

expandTypeSynonymAt :: Pos -> Type -> Converter Type

Applies expandTypeSynonym to the subterm at the given position.

If there are type constructor applications in parent positions, the type arguments from the parent positions are used to expand the type synonym as well.