Compiler.Haskell.Subterm

Description

This module contains data types and function for working with subterms of Haskell expressions.

Positions

data Pos

Describes a position of a subterm within a Haskell expression.

Constructors

Pos [Int]

rootPos :: Pos

Position of the root expression.

pos :: Expr -> [Pos]

Gets all valid positions of subterms within the given Haskell expression.

above :: Pos -> Pos -> Bool

below :: Pos -> Pos -> Bool

leftOf :: Pos -> Pos -> Bool

rightOf :: Pos -> Pos -> Bool

Subterms

selectSubterm :: Expr -> Pos -> Maybe Expr

Selects a subterm of the given expression at the specified position.

Returns Nothing if there is no such subterm.

replaceSubterm

Arguments

:: Expr
The expression whose subterm to replace.

-> Pos
The position of the subterm.

-> Expr
The expression to replace the subterm with.

-> Maybe Expr

Replaces a subterm of the given expression at the specified position with another expression.

Returns Nothing if there is no such subterm.

replaceSubterms :: Expr -> [(Pos, Expr)] -> Maybe Expr

Replaces all subterms at the given positions with other expressions.

Returns Nothing if any of the subterms could not be replaced

Searching for subterms

findSubtermPos :: (Expr -> Bool) -> Expr -> [Pos]

Gets a list of positions for subterms of the given expression that satisfy the provided predicate.

findSubterms :: (Expr -> Bool) -> Expr -> [Expr]

Gets a list of subterms of the given expression that satisfy the provided predicate.

Bound variables

boundVarsAt :: Expr -> Pos -> Set QName

Gets the names of variables that are bound by lambda abstractions or variable patterns in case-expressions at the given position of an expression.

Returns the empty set if the position is invalid.

usedVarsAt :: Expr -> Pos -> Set QName

Gets the names of variables that are used by the subterm t the given position.

The returned set also contains function and argument names. To get a set of variable names only, intersect the result with boundVarsAt.

Returns the empty set if the position is invalid.