Compiler.Haskell.Unification

Description

This module contains functions for calculating the most general unificator (mgu) for Haskell type expressions.

Documentation

unify :: Type -> Type -> Converter (Subst Type)

Calculates the mgu of the given type expressions.

The algorithm will preferably map the internal variable names to non-internal variables. This ensures that the names specified by the user are preserved. Otherwise variables in the first argument preferably mapped to variables in the second argument.

Reports a fatal error if the types cannot be unified.

unifyAll :: [Type] -> Converter (Subst Type)

Computes the most general unificator for all given type expressions.