Compiler.IR.Subst

Description

This module contains a definition of substitutions for expressions and type expressions.

Substitutions

data Subst a

A substitution is a mapping from Haskell variable names to expressions (i.e. Subst Expr) or type expressions (i.e. Subst Type).

When the substitution is applied (see applySubst) the source span of the substituted variable can be inserted into the (type) expression it is replaced by (e.g. to rename a variable without loosing source spans).

Construction

identitySubst :: Subst a

A substitution that does not change an expression.

singleSubst :: QName -> a -> Subst a

Creates a new substitution that maps the variable with the given name to the given expression or type expression.

singleSubst' :: QName -> (SrcSpan -> a) -> Subst a

Like singleSubst, but can be used to preserve the source span of the variable replaced by applySubst.

Composition

composeSubst :: ApplySubst a a => Subst a -> Subst a -> Subst a

Creates a new substitution that applies both given substitutions after each other.

composeSubsts :: ApplySubst a a => [Subst a] -> Subst a

Creates a new substitution that applies all given substitutions after each other.

Modification

filterSubst :: (QName -> Bool) -> Subst a -> Subst a

Creates a new substitution whose domain does not contain the names that match the given predicate.

substWithout :: Subst a -> [QName] -> Subst a

Creates a new substitution whose domain does not contain the given names.

Application

class ApplySubst a b where

Type class for applying a substitution that replaces variables by values of type a on values of type b.

Methods

applySubst :: Subst a -> b -> b

Rename arguments

renameTypeArgsSubst :: [TypeVarDecl] -> Converter ([TypeVarDecl], Subst Type)

Creates a substitution that renames the given type variables to fresh variables.

Returns the new names for the type variables and the substitution.

renameTypeArgs :: ApplySubst Type a => [TypeVarDecl] -> a -> Converter ([TypeVarDecl], a)

Renames the given type variables in the given expression or type to fresh type variables.

renameArgsSubst :: [VarPat] -> Converter ([VarPat], Subst Expr)

Creates a substitution that renames the arguments bound by the given variable patterns to fresh variables.

Returns the new names for the variables and the substitution.

renameArgs :: ApplySubst Expr a => [VarPat] -> a -> Converter ([VarPat], a)

Renames the arguments bound by the given variable patterns in the given expression to fresh variables.

Returns the new names for the variables and the resulting expression.