Compiler.Monad.Class.Hoistable

Description

This module defines type classes for converting between monads and their monad transformer counterparts.

Documentation

class MonadTrans t => Hoistable t where

Type class for monad transformers whose inner monad can be lifted from Identity to some arbitry Monad.

Methods

hoist :: Monad m => t Identity a -> t m a

Lifts the transformed identity monad to any transformed monad.

class Hoistable t => UnHoistable t where

Type class for Hoistable types whose hoist method can be undone.

Methods

unhoist :: Monad m => t m a -> m (t Identity a)

Undoes hoist.

hoistMaybe :: Monad m => Maybe a -> MaybeT m a

Lifts a Maybe value to an arbitrary monad.

Since Maybe and MaybeT Identity are distinct types, no equivalent Hoistable instance can be specified.