Compiler.Monad.Converter

Description

This module a state monad which allows the compiler's state (see Comiler.Environment) to be passed implicitly throught the converter.

There are also utility functions to modify the state and retreive information stored in the state.

State monad

type Converter = ConverterT Identity

Type synonym for the state monad used by the converter.

All converter functions usually require the current Environment to perform the conversion. This monad allows these functions to pass the environment around implicitly.

Additionally the converter can report error messages and warnings to the user if there is a problem while converting.

runConverter :: Converter a -> Environment -> Reporter (a, Environment)

Runs the converter with the given initial environment and returns the converter's result as well as the final environment.

evalConverter :: Converter a -> Environment -> Reporter a

Runs the converter with the given initial environment and returns the converter's result.

execConverter :: Converter a -> Environment -> Reporter Environment

Runs the converter with the given initial environment and returns the final environment.

State monad transformer

data ConverterT m a

A state monad used by the converter parametrerized by the inner monad m.

runConverterT :: Monad m => ConverterT m a -> Environment -> ReporterT m (a, Environment)

Runs the converter with the given initial environment and returns the converter's result as well as the final environment.

evalConverterT :: Monad m => ConverterT m a -> Environment -> ReporterT m a

Runs the converter with the given initial environment and returns the converter's result.

execConverterT :: Monad m => ConverterT m a -> Environment -> ReporterT m Environment

Runs the converter with the given initial environment and returns the final environment.

lift :: (MonadTrans t, Monad m) => m a -> t m a

Lift a computation from the argument monad to the constructed monad.

lift' :: Monad m => ReporterT m a -> ConverterT m a

Converts a reporter to a converter.

hoist :: (Hoistable t, Monad m) => t Identity a -> t m a

Lifts the transformed identity monad to any transformed monad.

Using IO actions in converters

type ConverterIO = ConverterT IO

A converter with an IO action as its inner monad.

Modifying environments

getEnv :: Monad m => ConverterT m Environment

Gets the current environment.

inEnv :: Monad m => (Environment -> a) -> ConverterT m a

Gets a specific component of the current environment using the given function to extract the value from the environment.

putEnv :: Monad m => Environment -> ConverterT m ()

Sets the current environment.

modifyEnv :: Monad m => (Environment -> Environment) -> ConverterT m ()

Applies the given function to the environment.

modifyEnv' :: Monad m => (Environment -> (a, Environment)) -> ConverterT m a

Gets a specific component and modifies the environment.

Encapsulating environments

localEnv :: Monad m => ConverterT m a -> ConverterT m a

Runs the given converter and returns its result but discards all modifications to the environment.

moduleEnv :: Monad m => ConverterT m (a, ModuleInterface) -> ConverterT m a

Runs the given converter and returns its result.

Like localEnv, the modifications to the environment are discared but the returned ModuleInterface is inserted into the environment.

sectionEnv :: Monad m => ConverterT m a -> ConverterT m a

Runs the given converter with envInSection set to True.

shadowVarPats :: Monad m => [VarPat] -> ConverterT m a -> ConverterT m a

Adds entries for variable patterns during the execution of the given converter.

Unlike localEnv, all modifications to the environment are keept (except for added entries), except for the definition of the variables.