Compiler.Pass.TypeInferencePass

Description

This module contains a compiler pass that infers the types of all function declarations including the types of the sub-expressions of their right-hand sides and type arguments to pass to used functions and constructors.

TODO write documentation and examples

Documentation

typeInferencePass :: DependencyAwarePass FuncDecl

A compiler pass that infers the types of (mutually recursive) function declarations in the given strongly connected component of the function dependency graph.