Compiler.Pipeline

Description

This module contains the compiler pipeline for the translation of the intermediate representation that is generated by the front end into a format that is accepted by the back end.

The compiler pipeline is organized into compiler Passes. Each pass performs some transformation on the converted module.

Documentation

pipeline :: [Pass Module]

The passes of the compiler pipeline.

runPipeline :: Module -> Converter Module

Runs the compiler pipeline on the given module.