Compiler.Converter.FuncDecl.Common

Description

This function contains auxilary functions that are used both to translate recursive and non-recursive Haskell functions to Coq.

Type inference

inferAndInsertTypeSigs :: DependencyComponent FuncDecl -> Converter (DependencyComponent FuncDecl)

Infers the types of the function declarations in the given strongly connected component and inserts type signatures into the environment.

Returns the function declarations where the type arguments of all function and constructor applications on the right-hand side are applied visibly.

splitFuncType

Arguments

:: QName
The name of the function to display in error messages.

-> [VarPat]
The argument variable patterns whose types to split of.

-> Type
The type to split.

-> Converter ([Type], Type)

Splits the annotated type of a Haskell function with the given arguments into its argument and return types.

Type synonyms are expanded if neccessary.

Function environment entries

defineFuncDecl :: FuncDecl -> Converter ()

Inserts the given function declaration into the current environment.

Code generation

convertFuncHead

Arguments

:: QName
The name of the function.

-> [VarPat]
The function argument patterns.

-> Converter (Qualid, [Binder], Maybe Term)

Converts the name, arguments and return type of a function to Coq.

This code is shared between the conversion functions for recursive and no recursive functions (see convertNonRecFuncDecl and convertRecFuncDecls).