Compiler.Converter.QuickCheck

Description

This module contains functions for converting simple QuickCheck properties to Coq Theorem sentences.

QuickCheck properties are functions that start with prop_ by convection. However, the compiler uses the type of the property to determine whether it is a QuickCheck property or not (i.e. all QuickCheck properties must return a value of type Property).

Documentation

quickCheckModuleName :: ModName

The name of the QuickCheck module.

quickCheckPropertyName :: QName

The name of the Property data type.

isQuickCheckProperty :: FuncDecl -> Converter Bool

Tests whether the given declaration is a QuickCheck property.

QuickCheck properties are functions with the prefix prop_ or which return a value of type Property.

containsQuickCheckProperty :: DependencyComponent FuncDecl -> Converter Bool

Tests whether the given strongly connected component of the function dependency graph contains a QuickCheck property.

filterQuickCheckProperties :: [DependencyComponent FuncDecl] -> Converter ([FuncDecl], [DependencyComponent FuncDecl])

Partitions the given list of strongly connected components of the function dependency graph into a list of QuickCheck properties and dependency components that contain no QuickCheck properties.

Reports a fatal error message if there is there is a (mutually) recursive QuickCheck property in one of the components.

convertQuickCheckProperty :: FuncDecl -> Converter [Sentence]

Converts the given QuickCheck property to a Coq Theorem with an empty Proof.