Compiler.Haskell.AST

Description

An alternative AST that represents the subset of Haskell modules that comply with our restrictions on the accepted input format.

The aim of this module is to make the conversion functions easier to comprehend and reduce coupling with the haskell-src-ext package.

This AST cannot be parsed directly. Instances of the AST are usually created by converting an existing AST (e.g. from the haskell-src-ext package). For testing purposes instances may be created directly.

Documentation

type TypeVarIdent = String

The name of a type variable.

data Name

An identifier or a symbolic name.

The constructors of this type do not contain source spans because Names are intended to be comparable. They are used as keys to identify nodes of the dependency graph for example.

Constructors

Ident String
An identifier, e.g. Ident "f" for a function f.

Symbol String
A symbolic name, e.g. Symbol "+" for (+).

data QName

A potentially qualified Name.

Constructors

Qual ModName Name
A qualified Name.

UnQual Name
An unqualified Name.

type ModName = String

The name of a module.

type VarName = QName

The name of a function or build-in operator used in prefix notation, e.g. f x y or (+) n m

type ConName = QName

The name of an constructor used in prefix notation, e.g. (:) x xs.

type TypeConName = QName

The name of a type or type constructor, e.g. Integer or [] a

data DeclIdent

The name of a function, data type, type synonym or constructor defined by the user including location information.

Because the user cannot declare symbols at the moment, the constructor of this data type takes a String and not a Name.

Constructors

DeclIdent SrcSpan String

type TypeVarDecl = DeclIdent

The name of a type variable declaration in the head of a data type or type synonym declaration including location information.

data ConPat

A constructor pattern used in an alternative of a case expression.

The only purpose of this data type is to add location information to a ConName.

Constructors

ConPat SrcSpan ConName

data VarPat

A variable pattern used as an argument to a function, lambda abstraction or constructor pattern.

The only purpose of this data type is to add location information to the identifer for a variable.

Constructors

VarPat SrcSpan String

internalIdentChar :: Char

The character that is used to mark internal identifiers.

This is used for example to generate fresh identifiers that don't conflict with actual Haskell identifiers or to hide entries in module interfaces.

isInternalIdent :: String -> Bool

Tests whether the given Haskell identifier was generated for internal use only (i.e., contains internalIdentChar).

isInternalName :: Name -> Bool

Tests whether the given Haskell name was generated for interal use only (i.e., it is an identifier that matches isInternalIdent).

isInternalQName :: QName -> Bool

Tests whether the given qualified Hasell name was generted for internal use only (i.e., the name or module name are internal according to isInternalIdent and isInternalName, respectively).

data Module

A module declaration.

If the module has no module header, the module name defauts to Main.

Constructors

Module

Fields

modSrcSpan :: SrcSpan

modName :: ModName

modImports :: [ImportDecl]

modTypeDecls :: [TypeDecl]

modTypeSigs :: [TypeSig]

modPragmas :: [Pragma]

modFuncDecls :: [FuncDecl]

data Comment

A comment.

Comments are collected during parsing. However, the final AST contains no comments. Pragmas (see DecArgPragma) are extracted from comments during simplification.

Constructors

BlockComment SrcSpan String
A multi-line comment (i.e., {- ... -}).

LineComment SrcSpan String
A single-line comment (i.e., -- ...).

customPragmaPrefix :: String

All custom pragmas of the compiler start with HASKELL_TO_COQ.

data Pragma

Data type for custom {--} pragmas.

Constructors

DecArgPragma SrcSpan String String
A {--} pragma.

prettyPragma :: Doc -> Doc

Pretty prints a custom {--} pragma with the given contents.

data ImportDecl

An import declaration.

Constructors

ImportDecl

Fields

importSrcSpan :: SrcSpan

importName :: ModName

data TypeDecl

A data type or type synonym declaration.

While it is allowed to define constructors in infix notation, data type and type synonym declarations must not be in infix notation. This is because the TypeOperators language extension is not supported.

Constructors

DataDecl SrcSpan DeclIdent [TypeVarDecl] [ConDecl]

TypeSynDecl SrcSpan DeclIdent [TypeVarDecl] Type

data TypeSig

A type signature of one or more function declarations.

Constructors

TypeSig SrcSpan [DeclIdent] Type

data FuncDecl

A function declaration.

Even though there is not a separate constructor, it is allowed to define functions in infix notation e.g.:

    append :: [a] -> [a] -> [a]
    xs `append` ys = ...
  

But it is currently not supported to use symbolic names for such definitions. Thus the following is not allowed:

    (++) :: [a] -> [a] -> [a]
    xs ++ ys = ...
  

The precedence and associativity (fixity) of infix declarations can be specified as well.

    infixr 5 `append`
  

Constructors

FuncDecl SrcSpan DeclIdent [VarPat] Expr

data ConDecl

A constructor declaration.

Even though there is not a dedicated constructor, the constructor is allowed to be in infix notation, but the name of the constructor must not be a symbol.

Constructors

ConDecl

Fields

SrcSpan

A source span that spans the entire constructor declaration.

DeclIdent

The name of the constructor.

[Type]

The types of the constructor arguments.

data TypeSchema

A Haskell type expression with explicitly introduced type variables.

Constructors

TypeSchema SrcSpan [TypeVarDecl] Type

data Type

A Haskell type expression.

Build-in types are represented by applications of their type constructors. E.g. the type [a] is represented as TypeApp (TypeCon "[]") (TypeVar "a"). The only exception to this rule is the function type a -> b. It is represented directly as TypeFunc (TypeVar "a") (TypeVar "b"). The syntax (->) a b is not supported at the moment. This is due to the special role of functions during the translation to Coq.

Constructors

TypeVar SrcSpan TypeVarIdent
A type variable.

TypeCon SrcSpan TypeConName
A type constructor.

TypeApp SrcSpan Type Type
A type constructor application.

TypeFunc SrcSpan Type Type
A function type.

data Expr

A Haskell expression.

Even though there are no dedicated constructors, the infix applications of functions and constructors (including left and right sections) are supported. This kind of syntactic sugar is removed during simplification (see Compiler.Haskell.Simplifier).

Constructors

Con SrcSpan ConName
A constructor.

Var SrcSpan VarName
A function or local variable.

App SrcSpan Expr Expr
Function or constructor application.

If SrcSpan Expr Expr Expr
if expression.

Case SrcSpan Expr [Alt]
case expression.

Undefined SrcSpan
Error term undefined.

ErrorExpr SrcSpan String
Error term error "message".

IntLiteral SrcSpan Integer
An integer literal.

Lambda SrcSpan [VarPat] Expr
A lambda abstraction.

ExprTypeSig SrcSpan Expr TypeSchema
A type annotation.

data Alt

One alternative of a case expression.

Every alternative of a case expression matches a constructor of the matched expression's data type. All arguments of the constructor pattern are variable patterns.

Constructors

Alt

Fields

SrcSpan

A source span that spans the entire alternative.

ConPat

The name of the constructor matched by this alternative.

[VarPat]

Variable patterns for the arguments of the constructor.

Expr

The right hand side of this alternative.

fromDeclIdent :: DeclIdent -> String

Extracts the actual identifier from an identifier in a declaration.

fromVarPat :: VarPat -> String

Extracts the actual identifier from a variable pattern.

identFromName :: Name -> Maybe String

Extracts an identifier from a Haskell name. Returns Nothing if the given name is a symbol and not an identifier.

identFromQName :: QName -> Maybe String

Extracts an identifier from an unqualified Haskell name.

Returns Nothing if the given name is qualified or a symbol and not an identifier.

splitType :: Type -> Int -> ([Maybe Type], Maybe Type)

Splits the type of a function or constructor with the given arity into the argument and return types.

class GetDeclIdent decl where

This type class provides a getter for the name of declarations of type decl.

Methods

getDeclIdent :: decl -> DeclIdent

prettyDeclIdents :: GetDeclIdent decl => [decl] -> String

Gets the names of the given declarations and concatenates them with commas.

class GetSrcSpan a where

This type class provides a getter for the source span of an AST node of type a.

Methods

getSrcSpan :: a -> SrcSpan

Gets the source span of the given AST node.

funcType :: SrcSpan -> [Type] -> Type -> Type

Creates a function type with the given argument and return types.

typeApp

Arguments

:: SrcSpan
The source span to insert into generated nodes.

-> TypeConName
The name of the type constructor to apply.

-> [Type]
The type arguments to pass to the type constructor.

-> Type

Creates a type constructor application type.

The given source span is inserted into the generated type constructor and every generated type constructor application.

app :: SrcSpan -> Expr -> [Expr] -> Expr

Creates an expression for applying the given expression to the provided arguments.

The given source span is inserted into the generated function application expressions.

varApp

Arguments

:: SrcSpan
The source span to insert into generated nodes.

-> VarName
The name of the function to apply.

-> [Expr]
The arguments to pass to the function.

-> Expr

Creates an expression for applying the function with the given name.

The given source span is inserted into the generated function reference and every generated function application.

conApp

Arguments

:: SrcSpan
The source span to insert into generated nodes.

-> ConName
The name of the constructor to apply.

-> [Expr]
The arguments to pass to the constructor.

-> Expr

Creates a data constructor application expression.

The given source span is inserted into the generated constructor reference and every generated constructor application.

preludeModuleName :: ModName

The name of the Prelude module.

TODO once import ... as ... is supported, the Prelude could be renamed by the user.

unitTypeConName :: TypeConName

The name of the unit type constructor.

tupleTypeConName :: Int -> TypeConName

The name of the n-ary tuple type constructor.

listTypeConName :: TypeConName

The name of the list type constructor.

unitConName :: ConName

Name of the unit data constructor.

nilConName :: ConName

The name of the empty list data constructor.

consConName :: ConName

The name of the non empty list data constructor.

tupleConName :: Int -> ConName

The name of the n-ary tuple data constructor.

boolTypeConName :: TypeConName

When translating if expressions, we annotate the type of the condition with Bool. Because we do not support qualified identifiers we need to use this special symbol to prevent the user from shadowing Bool accidentaly with a custom function or local variable.

trueConName :: ConName

When translating boolean expressions in QuickCheck properties, we have to generate a check whether the result is True. Because we do not support qualified identifiers we need to use this special symbol to prevent the user from shadowing True accidentaly with a custom constructor.

negateOpName :: VarName

The unary prefix operator - is translated to the application of the negate function. Because we do not support qualified identifiers we need to use this special symbol to prevent the user from shadowing negate accidentaly with a custom function or local variable.

errorFuncName :: VarName

The name of the error function.

undefinedFuncName :: VarName

The name of the undefined function.